Simplify action handling, step 1, SaltServerActionService
# Conflicts:
#	java/code/src/com/redhat/rhn/domain/action/ActionFactory.java
#	java/code/src/com/suse/manager/webui/services/SaltServerActionService.java
