+++ java/code/src/com/redhat/rhn/domain/action/Action_legacyUser.hbm.xml
@@ -166,7 +166,7 @@ PUBLIC "-//Hibernate/Hibernate Mapping DTD 3.0//EN"
                 </subclass>
                 <!-- ConfigAction subclass: 16, 17, 18, 22 -->
                 <subclass name="com.redhat.rhn.domain.action.config.ConfigAction"
-                        discriminator-value="16" lazy="true">
+                        discriminator-value="-16" lazy="true">
                         <set lazy="true" cascade="all" inverse="true" name="configRevisionActions" >
                                 <key column="action_id" />
                                 <one-to-many
@@ -175,10 +175,14 @@ PUBLIC "-//Hibernate/Hibernate Mapping DTD 3.0//EN"
                         </set>
                         <subclass
                                 name="com.redhat.rhn.domain.action.config.ConfigDeployAction"
-                                discriminator-value="17" lazy="true">
+                                discriminator-value="16" lazy="true">
                         </subclass>
                         <subclass
                                 name="com.redhat.rhn.domain.action.config.ConfigVerifyAction"
+                                discriminator-value="17" lazy="true">
+                        </subclass>
+                        <subclass
+                                name="com.redhat.rhn.domain.action.config.ConfigDiffAction"
                                 discriminator-value="18" lazy="true">
                         </subclass>
                         <!-- TODO: These last two subclasses probably don't belong here.
