+++ java/code/src/com/redhat/rhn/domain/action/HardwareRefreshAction.java
@@ -15,9 +15,54 @@
 package com.redhat.rhn.domain.action;
 
 
+import com.redhat.rhn.domain.server.MinionSummary;
+
+import com.suse.manager.reactor.messaging.ApplyStatesEventMessage;
+import com.suse.salt.netapi.calls.LocalCall;
+import com.suse.salt.netapi.calls.modules.State;
+
+import java.util.Arrays;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+import java.util.stream.Collectors;
+
 /**
  * HardwareRefreshAction
  */
 public class HardwareRefreshAction extends Action {
+
+    /**
+     * @param minionSummaries a list of minion summaries of the minions involved in the given Action
+     * @return minion summaries grouped by local call
+     */
+    @Override
+    public Map<LocalCall<?>, List<MinionSummary>> getSaltCalls(List<MinionSummary> minionSummaries) {
+        Map<LocalCall<?>, List<MinionSummary>> ret = new HashMap<>();
+
+        // salt-ssh minions in the 'true' partition
+        // regular minions in the 'false' partition
+        Map<Boolean, List<MinionSummary>> partitionBySSHPush = minionSummaries.stream()
+                .collect(Collectors.partitioningBy(MinionSummary::isSshPush));
+
+        // Separate SSH push minions from regular minions to apply different states
+        List<MinionSummary> sshPushMinions = partitionBySSHPush.get(true);
+        List<MinionSummary> regularMinions = partitionBySSHPush.get(false);
+
+        if (!sshPushMinions.isEmpty()) {
+            ret.put(State.apply(List.of(
+                            ApplyStatesEventMessage.HARDWARE_PROFILE_UPDATE),
+                    Optional.empty()), sshPushMinions);
+        }
+        if (!regularMinions.isEmpty()) {
+            ret.put(State.apply(Arrays.asList(
+                            ApplyStatesEventMessage.SYNC_ALL,
+                            ApplyStatesEventMessage.HARDWARE_PROFILE_UPDATE),
+                    Optional.empty()), regularMinions);
+        }
+
+        return ret;
+    }
 
 }
