+++ java/code/src/com/redhat/rhn/domain/action/appstream/AppStreamAction.java
@@ -14,16 +14,28 @@
  */
 package com.redhat.rhn.domain.action.appstream;
 
+import static java.util.Collections.singletonList;
+
 import com.redhat.rhn.domain.action.Action;
+import com.redhat.rhn.domain.server.MinionSummary;
+
+import com.suse.manager.webui.services.SaltParameters;
+import com.suse.salt.netapi.calls.LocalCall;
+import com.suse.salt.netapi.calls.modules.State;
 
 import org.apache.commons.lang3.builder.EqualsBuilder;
 import org.apache.commons.lang3.builder.HashCodeBuilder;
 
+import java.util.Arrays;
+import java.util.HashMap;
 import java.util.HashSet;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
 import java.util.Set;
+import java.util.stream.Collectors;
 
 public class AppStreamAction extends Action {
-
     private static final long serialVersionUID = 1L;
     private Set<AppStreamActionDetails> details = new HashSet<>();
 
@@ -61,4 +73,33 @@ public class AppStreamAction extends Action {
                 .append(details)
                 .toHashCode();
     }
+
+
+
+    /**
+     * @param minionSummaries a list of minion summaries of the minions involved in the given Action
+     * @return minion summaries grouped by local call
+     */
+    @Override
+    public Map<LocalCall<?>, List<MinionSummary>> getSaltCalls(List<MinionSummary> minionSummaries) {
+        Map<LocalCall<?>, List<MinionSummary>> ret = new HashMap<>();
+
+        Map<Boolean, Set<AppStreamActionDetails>> det = getDetails().stream()
+                .collect(Collectors.partitioningBy(AppStreamActionDetails::isEnable, Collectors.toSet()));
+
+        var enableParams = det.get(true).stream()
+                .map(d -> d.getStream() == null ?
+                        singletonList(d.getModuleName()) :
+                        Arrays.asList(d.getModuleName(), d.getStream()))
+                .toList();
+        var disableParams = det.get(false).stream().map(AppStreamActionDetails::getModuleName).toList();
+
+        Optional<Map<String, Object>> params = Optional.of(Map.of(
+                SaltParameters.PARAM_APPSTREAMS_ENABLE, enableParams,
+                SaltParameters.PARAM_APPSTREAMS_DISABLE, disableParams
+        ));
+        ret.put(State.apply(List.of(SaltParameters.APPSTREAMS_CONFIGURE), params), minionSummaries);
+        return ret;
+    }
+
 }
