+++ java/code/src/com/redhat/rhn/domain/action/channel/SubscribeChannelsAction.java
@@ -15,13 +15,32 @@
 
 package com.redhat.rhn.domain.action.channel;
 
+import com.redhat.rhn.GlobalInstanceHolder;
 import com.redhat.rhn.domain.action.Action;
+import com.redhat.rhn.domain.server.MinionServer;
+import com.redhat.rhn.domain.server.MinionServerFactory;
+import com.redhat.rhn.domain.server.MinionSummary;
+import com.redhat.rhn.domain.server.ServerFactory;
+import com.redhat.rhn.domain.server.ServerGroupFactory;
+import com.redhat.rhn.manager.system.SystemManager;
+
+import com.suse.manager.reactor.messaging.ApplyStatesEventMessage;
+import com.suse.manager.webui.services.iface.SaltApi;
+import com.suse.salt.netapi.calls.LocalCall;
+import com.suse.salt.netapi.calls.modules.State;
+
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+import java.util.stream.Collectors;
 
 /**
  * SubscribeChannelsAction - Class representing a channel(s) subscription action
  */
 public class SubscribeChannelsAction extends Action {
 
+    private SaltApi saltApi = null;
     private SubscribeChannelsActionDetails details;
 
     /**
@@ -36,5 +55,41 @@ public class SubscribeChannelsAction extends Action {
      */
     public void setDetails(SubscribeChannelsActionDetails actionDetails) {
         this.details = actionDetails;
+    }
+
+    /**
+     * @param saltApiIn SaltApi to set
+     */
+    public void setSaltApi(SaltApi saltApiIn) {
+        this.saltApi = saltApiIn;
+    }
+
+    /**
+     * @param minionSummaries a list of minion summaries of the minions involved in the given Action
+     * @return minion summaries grouped by local call
+     */
+    @Override
+    public Map<LocalCall<?>, List<MinionSummary>> getSaltCalls(List<MinionSummary> minionSummaries) {
+
+        Map<LocalCall<?>, List<MinionSummary>> ret = new HashMap<>();
+        SystemManager sysMgr = new SystemManager(ServerFactory.SINGLETON, ServerGroupFactory.SINGLETON,
+                (null == saltApi) ? GlobalInstanceHolder.SALT_API : saltApi);
+
+        List<MinionServer> minions = MinionServerFactory.lookupByMinionIds(
+                minionSummaries.stream().map(MinionSummary::getMinionId).collect(Collectors.toSet()));
+
+        minions.forEach(minion ->
+                // change channels in DB and execult the ChannelsChangedEventMessageAction
+                // which regenerate pillar and refresh Tokens but does not execute a "state.apply channels"
+                sysMgr.updateServerChannels(
+                        details.getParentAction().getSchedulerUser(),
+                        minion,
+                        Optional.ofNullable(details.getBaseChannel()),
+                        details.getChannels())
+        );
+        ret.put(State.apply(List.of(ApplyStatesEventMessage.CHANNELS), Optional.empty()),
+                minionSummaries);
+
+        return ret;
     }
 }
