+++ java/code/src/com/redhat/rhn/domain/action/config/ConfigDeployAction.java
@@ -14,9 +14,62 @@
  */
 package com.redhat.rhn.domain.action.config;
 
+import com.redhat.rhn.domain.config.ConfigRevision;
+import com.redhat.rhn.domain.server.MinionSummary;
+
+import com.suse.manager.webui.services.ConfigChannelSaltManager;
+import com.suse.manager.webui.services.SaltParameters;
+import com.suse.salt.netapi.calls.LocalCall;
+import com.suse.salt.netapi.calls.modules.State;
+
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+import java.util.Set;
+import java.util.stream.Collectors;
+
 /**
  * ConfigDeployAction - Class representing TYPE_CONFIGFILES_DEPLOY
  */
 public class ConfigDeployAction extends ConfigAction {
+
+
+    /**
+     * Deploy files(files, directory, symlink) through state.apply
+     *
+     * @param minionSummaries a list of minion summaries of the minions involved in the given Action
+     * @return minion summaries grouped by local call
+     */
+    @Override
+    public Map<LocalCall<?>, List<MinionSummary>> getSaltCalls(List<MinionSummary> minionSummaries) {
+        Map<LocalCall<?>, List<MinionSummary>> ret = new HashMap<>();
+
+        Map<Long, MinionSummary> targetMap = minionSummaries.stream().
+                collect(Collectors.toMap(MinionSummary::getServerId, minionId-> minionId));
+
+        Map<MinionSummary, Set<ConfigRevision>> serverConfigMap = getConfigRevisionActions()
+                .stream()
+                .filter(cra -> targetMap.containsKey(cra.getServer().getId()))
+                .collect(Collectors.groupingBy(
+                        cra -> targetMap.get(cra.getServer().getId()),
+                        Collectors.mapping(ConfigRevisionAction::getConfigRevision, Collectors.toSet())));
+        Map<Set<ConfigRevision>, Set<MinionSummary>> revsServersMap = serverConfigMap.entrySet()
+                .stream()
+                .collect(Collectors.groupingBy(Map.Entry::getValue,
+                        Collectors.mapping(Map.Entry::getKey, Collectors.toSet())));
+        revsServersMap.forEach((configRevisions, selectedServers) -> {
+            List<Map<String, Object>> fileStates = configRevisions
+                    .stream()
+                    .map(revision -> ConfigChannelSaltManager.getInstance().getStateParameters(revision))
+                    .toList();
+            ret.put(State.apply(List.of(SaltParameters.CONFIG_DEPLOY_FILES),
+                            Optional.of(Collections.singletonMap(SaltParameters.PARAM_FILES, fileStates))),
+                    new ArrayList<>(selectedServers));
+        });
+        return ret;
+    }
 
 }
