+++ java/code/src/com/redhat/rhn/domain/action/rhnpackage/PackageLockAction.java
@@ -15,12 +15,29 @@
 
 package com.redhat.rhn.domain.action.rhnpackage;
 
+import static java.util.Collections.singletonMap;
+
+import com.redhat.rhn.common.db.datasource.DataResult;
 import com.redhat.rhn.domain.action.server.ServerAction;
+import com.redhat.rhn.domain.rhnpackage.PackageEvr;
+import com.redhat.rhn.domain.server.MinionSummary;
+import com.redhat.rhn.frontend.dto.PackageListItem;
 import com.redhat.rhn.manager.action.ActionManager;
 import com.redhat.rhn.manager.rhnpackage.PackageManager;
 
+import com.suse.manager.webui.services.SaltParameters;
+import com.suse.salt.netapi.calls.LocalCall;
+import com.suse.salt.netapi.calls.modules.State;
+
 import org.apache.logging.log4j.Level;
 import org.apache.logging.log4j.LogManager;
+
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
 
 /**
  * PackageLockAction
@@ -41,5 +58,29 @@ public class PackageLockAction extends PackageAction {
         }
 
         ActionManager.deleteActionsByIdAndType(this.getId(), this.getActionType().getId());
+    }
+
+    /**
+     * @param minionSummaries a list of minion summaries of the minions involved in the given Action
+     * @return minion summaries grouped by local call
+     */
+    @Override
+    public Map<LocalCall<?>, List<MinionSummary>> getSaltCalls(List<MinionSummary> minionSummaries) {
+        Map<LocalCall<?>, List<MinionSummary>> ret = new HashMap<>();
+
+        for (MinionSummary m : minionSummaries) {
+            DataResult<PackageListItem> setLockPkg = PackageManager.systemSetLockedPackages(
+                    m.getServerId(), getId(), null);
+            List<List<String>> pkgs = setLockPkg.stream().map(d -> Arrays.asList(d.getName(), d.getArch(),
+                    new PackageEvr(d.getEpoch(), d.getVersion(), d.getRelease(), d.getPackageType())
+                            .toUniversalEvrString())).toList();
+            LocalCall<Map<String, State.ApplyResult>> localCall =
+                    State.apply(List.of(SaltParameters.PACKAGES_PKGLOCK),
+                            Optional.of(singletonMap(SaltParameters.PARAM_PKGS, pkgs)));
+            List<MinionSummary> mSums = ret.getOrDefault(localCall, new ArrayList<>());
+            mSums.add(m);
+            ret.put(localCall, mSums);
+        }
+        return ret;
     }
 }
