+++ java/code/src/com/redhat/rhn/domain/action/rhnpackage/PackageUpdateAction.java
@@ -14,9 +14,79 @@
  */
 package com.redhat.rhn.domain.action.rhnpackage;
 
+import static java.util.Collections.singletonMap;
+import static java.util.stream.Collectors.groupingBy;
+import static java.util.stream.Collectors.joining;
+import static java.util.stream.Collectors.mapping;
+import static java.util.stream.Collectors.toList;
+
+import com.redhat.rhn.domain.errata.ErrataFactory;
+import com.redhat.rhn.domain.product.Tuple2;
+import com.redhat.rhn.domain.rhnpackage.PackageArch;
+import com.redhat.rhn.domain.rhnpackage.PackageEvr;
+import com.redhat.rhn.domain.rhnpackage.PackageName;
+import com.redhat.rhn.domain.server.MinionSummary;
+
+import com.suse.manager.webui.services.SaltParameters;
+import com.suse.salt.netapi.calls.LocalCall;
+import com.suse.salt.netapi.calls.modules.State;
+
+import java.util.Arrays;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+
 /**
  * PackageUpdateAction
  */
 public class PackageUpdateAction extends PackageAction {
+
+    /**
+     * @param minionSummaries a list of minion summaries of the minions involved in the given Action
+     * @return minion summaries grouped by local call
+     */
+    @Override
+    public Map<LocalCall<?>, List<MinionSummary>> getSaltCalls(List<MinionSummary> minionSummaries) {
+        Map<LocalCall<?>, List<MinionSummary>> ret = new HashMap<>();
+
+        List<Long> sids = minionSummaries.stream().map(MinionSummary::getServerId).collect(toList());
+
+        List<String> nevraStrings = getDetails().stream().map(details -> {
+            PackageName name = details.getPackageName();
+            PackageEvr evr = details.getEvr();
+            PackageArch arch = details.getArch();
+            return name.getName() + "-" + evr.toUniversalEvrString() + "." + arch.getLabel();
+        }).collect(toList());
+
+        List<Tuple2<Long, Long>> retractedPidSidPairs = ErrataFactory.retractedPackagesByNevra(nevraStrings, sids);
+        Map<Long, List<Long>> retractedPidsBySid = retractedPidSidPairs.stream()
+                .collect(groupingBy(Tuple2::getB, mapping(Tuple2::getA, toList())));
+        getServerActions().forEach(sa -> {
+            List<Long> packageIds = retractedPidsBySid.get(sa.getServerId());
+            if (packageIds != null) {
+                sa.fail("contains retracted packages: " +
+                        packageIds.stream().map(Object::toString).collect(joining(",")));
+            }
+        });
+        List<MinionSummary> filteredMinions = minionSummaries.stream()
+                .filter(ms -> retractedPidsBySid.get(ms.getServerId()) == null ||
+                        retractedPidsBySid.get(ms.getServerId()).isEmpty())
+                .collect(toList());
+
+        List<List<String>> pkgs =
+                getDetails().stream().map(d -> Arrays.asList(d.getPackageName().getName(),
+                        d.getArch().toUniversalArchString(), d.getEvr().toUniversalEvrString()))
+                .toList();
+        if (pkgs.isEmpty()) {
+            // Full system package update using update state
+            ret.put(State.apply(List.of(SaltParameters.PACKAGES_PKGUPDATE), Optional.empty()), filteredMinions);
+        }
+        else {
+            ret.put(State.apply(List.of(SaltParameters.PACKAGES_PKGINSTALL),
+                    Optional.of(singletonMap(SaltParameters.PARAM_PKGS, pkgs))), filteredMinions);
+        }
+        return ret;
+    }
 
 }
