+++ java/code/src/com/redhat/rhn/domain/action/salt/inspect/ImageInspectAction.java
@@ -16,6 +16,19 @@ package com.redhat.rhn.domain.action.salt.inspect;
 
 import com.redhat.rhn.domain.action.Action;
 import com.redhat.rhn.domain.action.ActionFormatter;
+import com.redhat.rhn.domain.image.ImageStore;
+import com.redhat.rhn.domain.image.ImageStoreFactory;
+import com.redhat.rhn.domain.server.MinionSummary;
+
+import com.suse.salt.netapi.calls.LocalCall;
+import com.suse.salt.netapi.calls.modules.State;
+
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.LinkedList;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
 
 /**
  * ImageInspectAction
@@ -49,6 +62,47 @@ public class ImageInspectAction extends Action {
             formatter = new ImageInspectActionFormatter(this);
         }
         return formatter;
+    }
+
+    /**
+     * @param minionSummaries a list of minion summaries of the minions involved in the given Action
+     * @return minion summaries grouped by local call
+     */
+    @Override
+    public Map<LocalCall<?>, List<MinionSummary>> getSaltCalls(List<MinionSummary> minionSummaries) {
+        if (details == null) {
+            return Collections.emptyMap();
+        }
+        return ImageStoreFactory.lookupById(details.getImageStoreId())
+                .map(store -> imageInspectAction(minionSummaries, store))
+                .orElseGet(Collections::emptyMap);
+    }
+
+    private Map<LocalCall<?>, List<MinionSummary>> imageInspectAction(List<MinionSummary> minions,
+                                                                      ImageStore store) {
+        Map<String, Object> pillar = new HashMap<>();
+        Map<LocalCall<?>, List<MinionSummary>> result = new HashMap<>();
+        if (ImageStoreFactory.TYPE_OS_IMAGE.equals(store.getStoreType())) {
+            pillar.put("build_id", "build" + details.getBuildActionId());
+            LocalCall<Map<String, State.ApplyResult>> apply = State.apply(
+                    Collections.singletonList("images.kiwi-image-inspect"),
+                    Optional.of(pillar));
+            result.put(apply, minions);
+            return result;
+        }
+        else {
+            List<ImageStore> imageStores = new LinkedList<>();
+            imageStores.add(store);
+            Map<String, Object> dockerRegistries = ImageStore.dockerRegPillar(imageStores);
+            pillar.put("docker-registries", dockerRegistries);
+            pillar.put("imagename", store.getUri() + "/" + details.getName() + ":" + details.getVersion());
+            pillar.put("build_id", "build" + details.getBuildActionId());
+            LocalCall<Map<String, State.ApplyResult>> apply = State.apply(
+                    Collections.singletonList("images.profileupdate"),
+                    Optional.of(pillar));
+            result.put(apply, minions);
+            return result;
+        }
     }
 
 }
