+++ java/code/src/com/redhat/rhn/domain/action/scap/ScapAction.java
@@ -14,12 +14,29 @@
  */
 package com.redhat.rhn.domain.action.scap;
 
+import static java.util.Collections.singletonList;
+import static java.util.Collections.singletonMap;
+import static java.util.stream.Collectors.toList;
+
 import com.redhat.rhn.common.localization.LocalizationService;
 import com.redhat.rhn.domain.action.Action;
+import com.redhat.rhn.domain.server.MinionSummary;
 import com.redhat.rhn.domain.server.Server;
 import com.redhat.rhn.domain.user.User;
 
+import com.suse.salt.netapi.calls.LocalCall;
+import com.suse.salt.netapi.calls.modules.State;
+
 import org.apache.commons.text.StringEscapeUtils;
+
+import java.util.Arrays;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+
 /**
  * ScapAction - Class representing TYPE_SCAP_*.
  */
@@ -64,4 +81,55 @@ public class ScapAction extends Action {
         return retval.toString();
     }
 
+    /**
+     * @param minionSummaries a list of minion summaries of the minions involved in the given Action
+     * @return minion summaries grouped by local call
+     */
+    @Override
+    public Map<LocalCall<?>, List<MinionSummary>> getSaltCalls(List<MinionSummary> minionSummaries) {
+
+       Map<LocalCall<?>, List<MinionSummary>> ret = new HashMap<>();
+        Map<String, Object> pillar = new HashMap<>();
+        Matcher profileMatcher = Pattern.compile("--profile (([\\w.-])+)")
+                .matcher(scapActionDetails.getParametersContents());
+        Matcher ruleMatcher = Pattern.compile("--rule (([\\w.-])+)")
+                .matcher(scapActionDetails.getParametersContents());
+        Matcher tailoringFileMatcher = Pattern.compile("--tailoring-file (([\\w./-])+)")
+                .matcher(scapActionDetails.getParametersContents());
+        Matcher tailoringIdMatcher = Pattern.compile("--tailoring-id (([\\w.-])+)")
+                .matcher(scapActionDetails.getParametersContents());
+
+        String oldParameters = "eval " +
+                scapActionDetails.getParametersContents() + " " + scapActionDetails.getPath();
+        pillar.put("old_parameters", oldParameters);
+
+        pillar.put("xccdffile", scapActionDetails.getPath());
+        if (scapActionDetails.getOvalfiles() != null) {
+            pillar.put("ovalfiles", Arrays.stream(scapActionDetails.getOvalfiles().split(","))
+                    .map(String::trim).collect(toList()));
+        }
+        if (profileMatcher.find()) {
+            pillar.put("profile", profileMatcher.group(1));
+        }
+        if (ruleMatcher.find()) {
+            pillar.put("rule", ruleMatcher.group(1));
+        }
+        if (tailoringFileMatcher.find()) {
+            pillar.put("tailoring_file", tailoringFileMatcher.group(1));
+        }
+        if (tailoringIdMatcher.find()) {
+            pillar.put("tailoring_id", tailoringIdMatcher.group(1));
+        }
+        if (scapActionDetails.getParametersContents().contains("--fetch-remote-resources")) {
+            pillar.put("fetch_remote_resources", true);
+        }
+        if (scapActionDetails.getParametersContents().contains("--remediate")) {
+            pillar.put("remediate", true);
+        }
+
+        ret.put(State.apply(singletonList("scap"),
+                        Optional.of(singletonMap("mgr_scap_params", (Object)pillar))),
+                minionSummaries);
+        return ret;
+    }
 }
